+++ unifi/network_firewall.go
@@ -2,6 +2,7 @@ package unifi
 
 import (
 	"context"
+	"fmt"
 
 	network "github.com/ubiquiti-community/go-unifi/client/network"
 
@@ -9,6 +10,10 @@ import (
 )
 
 func (c *ApiClient) ListFirewallPolicies(ctx context.Context, site uuid.UUID) ([]network.FirewallPolicy, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	return FetchAll(ctx, func(offset int32) (*network.FirewallPolicyPage, error) {
 		resp, err := c.network.client.GetFirewallPoliciesWithResponse(ctx, site, &network.GetFirewallPoliciesParams{
 			Offset: Ptr(offset),
@@ -22,6 +27,10 @@ func (c *ApiClient) ListFirewallPolicies(ctx context.Context, site uuid.UUID) ([
 }
 
 func (c *ApiClient) GetFirewallPolicy(ctx context.Context, site, id uuid.UUID) (*network.FirewallPolicy, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.GetFirewallPolicyWithResponse(ctx, site, id)
 	if err != nil {
 		return nil, err
@@ -30,11 +39,19 @@ func (c *ApiClient) GetFirewallPolicy(ctx context.Context, site, id uuid.UUID) (
 }
 
 func (c *ApiClient) DeleteFirewallPolicy(ctx context.Context, site, id uuid.UUID) error {
+	if c.network == nil {
+		return fmt.Errorf("Network API is unavailable")
+	}
+
 	_, err := c.network.client.DeleteFirewallPolicy(ctx, site, id)
 	return err
 }
 
 func (c *ApiClient) CreateFirewallPolicy(ctx context.Context, site uuid.UUID, data network.CreateFirewallPolicyJSONRequestBody) (*network.FirewallPolicy, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.CreateFirewallPolicyWithResponse(ctx, site, data)
 	if err != nil {
 		return nil, err
@@ -43,6 +60,10 @@ func (c *ApiClient) CreateFirewallPolicy(ctx context.Context, site uuid.UUID, da
 }
 
 func (c *ApiClient) UpdateFirewallPolicy(ctx context.Context, site, id uuid.UUID, data network.UpdateFirewallPolicyJSONRequestBody) (*network.FirewallPolicy, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.UpdateFirewallPolicyWithResponse(ctx, site, id, data)
 	if err != nil {
 		return nil, err
@@ -51,6 +72,10 @@ func (c *ApiClient) UpdateFirewallPolicy(ctx context.Context, site, id uuid.UUID
 }
 
 func (c *ApiClient) ListFirewallZones(ctx context.Context, site uuid.UUID) ([]network.FirewallZone, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	return FetchAll(ctx, func(offset int32) (*network.FirewallZonesPage, error) {
 		resp, err := c.network.client.GetFirewallZonesWithResponse(ctx, site, &network.GetFirewallZonesParams{
 			Offset: Ptr(offset),
@@ -64,6 +89,10 @@ func (c *ApiClient) ListFirewallZones(ctx context.Context, site uuid.UUID) ([]ne
 }
 
 func (c *ApiClient) GetFirewallZone(ctx context.Context, site, id uuid.UUID) (*network.FirewallZone, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.GetFirewallZoneWithResponse(ctx, site, id)
 	if err != nil {
 		return nil, err
@@ -72,11 +101,19 @@ func (c *ApiClient) GetFirewallZone(ctx context.Context, site, id uuid.UUID) (*n
 }
 
 func (c *ApiClient) DeleteFirewallZone(ctx context.Context, site, id uuid.UUID) error {
+	if c.network == nil {
+		return fmt.Errorf("Network API is unavailable")
+	}
+
 	_, err := c.network.client.DeleteFirewallZone(ctx, site, id)
 	return err
 }
 
 func (c *ApiClient) CreateFirewallZone(ctx context.Context, site uuid.UUID, data network.CreateFirewallZoneJSONRequestBody) (*network.FirewallZone, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.CreateFirewallZoneWithResponse(ctx, site, data)
 	if err != nil {
 		return nil, err
@@ -85,6 +122,10 @@ func (c *ApiClient) CreateFirewallZone(ctx context.Context, site uuid.UUID, data
 }
 
 func (c *ApiClient) UpdateFirewallZone(ctx context.Context, site, id uuid.UUID, data network.UpdateFirewallZoneJSONRequestBody) (*network.FirewallZone, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.UpdateFirewallZoneWithResponse(ctx, site, id, data)
 	if err != nil {
 		return nil, err
@@ -93,6 +134,10 @@ func (c *ApiClient) UpdateFirewallZone(ctx context.Context, site, id uuid.UUID, 
 }
 
 func (c *ApiClient) GetFirewallPolicyOrdering(ctx context.Context, site uuid.UUID, sourceZone, destinationZone uuid.UUID) (*network.OrderedFirewallPolicyIDs, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.GetFirewallPolicyOrderingWithResponse(ctx, site, &network.GetFirewallPolicyOrderingParams{
 		SourceFirewallZoneId:      sourceZone,
 		DestinationFirewallZoneId: destinationZone,
@@ -104,6 +149,10 @@ func (c *ApiClient) GetFirewallPolicyOrdering(ctx context.Context, site uuid.UUI
 }
 
 func (c *ApiClient) PatchFirewallPolicy(ctx context.Context, site, id uuid.UUID, data network.PatchFirewallPolicyJSONRequestBody) (*network.FirewallPolicy, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.PatchFirewallPolicyWithResponse(ctx, site, id, data)
 	if err != nil {
 		return nil, err
