+++ unifi/network_traffic_matching_list.go
@@ -2,6 +2,7 @@ package unifi
 
 import (
 	"context"
+	"fmt"
 
 	network "github.com/ubiquiti-community/go-unifi/client/network"
 
@@ -9,6 +10,10 @@ import (
 )
 
 func (c *ApiClient) ListTrafficMatchingLists(ctx context.Context, site uuid.UUID) ([]network.TrafficMatchingList, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	return FetchAll(ctx, func(offset int32) (*network.TrafficMatchingListsPage, error) {
 		resp, err := c.network.client.GetTrafficMatchingListsWithResponse(ctx, site, &network.GetTrafficMatchingListsParams{
 			Offset: Ptr(offset),
@@ -22,6 +27,10 @@ func (c *ApiClient) ListTrafficMatchingLists(ctx context.Context, site uuid.UUID
 }
 
 func (c *ApiClient) GetTrafficMatchingList(ctx context.Context, site, id uuid.UUID) (*network.TrafficMatchingList, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.GetTrafficMatchingListWithResponse(ctx, site, id)
 	if err != nil {
 		return nil, err
@@ -30,11 +39,19 @@ func (c *ApiClient) GetTrafficMatchingList(ctx context.Context, site, id uuid.UU
 }
 
 func (c *ApiClient) DeleteTrafficMatchingList(ctx context.Context, site, id uuid.UUID) error {
+	if c.network == nil {
+		return fmt.Errorf("Network API is unavailable")
+	}
+
 	_, err := c.network.client.DeleteTrafficMatchingList(ctx, site, id)
 	return err
 }
 
 func (c *ApiClient) CreateTrafficMatchingList(ctx context.Context, site uuid.UUID, data network.CreateTrafficMatchingListJSONRequestBody) (*network.TrafficMatchingList, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.CreateTrafficMatchingListWithResponse(ctx, site, data)
 	if err != nil {
 		return nil, err
@@ -43,6 +60,10 @@ func (c *ApiClient) CreateTrafficMatchingList(ctx context.Context, site uuid.UUI
 }
 
 func (c *ApiClient) UpdateTrafficMatchingList(ctx context.Context, site, id uuid.UUID, data network.UpdateTrafficMatchingListJSONRequestBody) (*network.TrafficMatchingList, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.UpdateTrafficMatchingListWithResponse(ctx, site, id, data)
 	if err != nil {
 		return nil, err
