+++ unifi/network_wifi.go
@@ -2,6 +2,7 @@ package unifi
 
 import (
 	"context"
+	"fmt"
 
 	network "github.com/ubiquiti-community/go-unifi/client/network"
 
@@ -9,6 +10,10 @@ import (
 )
 
 func (c *ApiClient) ListWifiBroadcasts(ctx context.Context, site uuid.UUID) ([]network.WifiBroadcastOverview, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	return FetchAll(ctx, func(offset int32) (*network.IntegrationWifiBroadcastPageDto, error) {
 		resp, err := c.network.client.GetWifiBroadcastPageWithResponse(ctx, site, &network.GetWifiBroadcastPageParams{
 			Offset: Ptr(offset),
@@ -22,6 +27,10 @@ func (c *ApiClient) ListWifiBroadcasts(ctx context.Context, site uuid.UUID) ([]n
 }
 
 func (c *ApiClient) GetWifiBroadcast(ctx context.Context, site, id uuid.UUID) (*network.WifiBroadcastDetails, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.GetWifiBroadcastDetailsWithResponse(ctx, site, id)
 	if err != nil {
 		return nil, err
@@ -30,11 +39,19 @@ func (c *ApiClient) GetWifiBroadcast(ctx context.Context, site, id uuid.UUID) (*
 }
 
 func (c *ApiClient) DeleteWifiBroadcast(ctx context.Context, site, id uuid.UUID) error {
+	if c.network == nil {
+		return fmt.Errorf("Network API is unavailable")
+	}
+
 	_, err := c.network.client.DeleteWifiBroadcast(ctx, site, id, &network.DeleteWifiBroadcastParams{})
 	return err
 }
 
 func (c *ApiClient) CreateWifiBroadcast(ctx context.Context, site uuid.UUID, data network.CreateWifiBroadcastJSONRequestBody) (*network.WifiBroadcastDetails, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.CreateWifiBroadcastWithResponse(ctx, site, data)
 	if err != nil {
 		return nil, err
@@ -43,6 +60,10 @@ func (c *ApiClient) CreateWifiBroadcast(ctx context.Context, site uuid.UUID, dat
 }
 
 func (c *ApiClient) UpdateWifiBroadcast(ctx context.Context, site, id uuid.UUID, data network.UpdateWifiBroadcastJSONRequestBody) (*network.WifiBroadcastDetails, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	resp, err := c.network.client.UpdateWifiBroadcastWithResponse(ctx, site, id, data)
 	if err != nil {
 		return nil, err
